+++ TerrariaPack.cs
@@ -79,6 +79,7 @@ public sealed class Terraria : SimpleTCPPack
                 new Effect("Force mount", "force_mount", PlayerFolder) {Price = 50, Description = "Temporarily force the streamer to ride a mount"},
                 new Effect("Shoot bombs", "shoot_bombs", PlayerFolder) {Price = 600, Description = "Temporarily cause the streamer to drop and shoot bombs that are capable of destroying blocks"},
                 new Effect("Shoot grenades", "shoot_grenades", PlayerFolder) {Price = 200, Description = "Temporarily cause the streamer to drop and shoot grenades"},
+                new Effect("Increase knockback", "increase_knockback", PlayerFolder) {Price = 40, Description = "Temporarily increase the knockback received and dealt by the streamer"},
                 new Effect("Increase jump height", "jump_boost", PlayerFolder) {Price = 40, Description = "Temporarily increase the streamer's jump height by a large amount"},
                 new Effect("Increase movement speed", "run_boost", PlayerFolder) {Price = 40, Description = "Temporarily increase the streamer's movement speed by a large amount"},
                 new Effect("Slippery boots", "icy_feet", PlayerFolder) {Price = 40, Description = "Temporarily make the ground very slippery under the streamer"},