+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added monolith-related effects; which temporarily change the background in-game.
 - Added "Switch Loadout" effect.
 - Added "Hiccup" effect.
+- Added "Increase Knockback" effect.
 - Changed "Fart Effect" to spawn a poo item when the player is well-fed.
 - Changed Dungeon Guardian effects to hide the chat message if the player is in the dungeon pre-skeletron.
 - Changed "Money Boost" effect to only detect NPC strikes from the affected player.